Added: args validation

+++ options/bucket/clearBucket.js
@@ -1,7 +1,10 @@
 import { HOST } from '../../env.js'
+import { argsValidation } from '../../utilites/args-validation.js'
 
 export async function clearBucketOption() {
     const bucketName = process.argv[4]
+    if (!argsValidation([bucketName])) { return }
+
     const res = await fetch(`${HOST}/bucket/clear?bucketName=${bucketName}`, {
         method: "DELETE",
     })
