+++ options/bucket/deleteBucketOption.js
@@ -1,8 +1,10 @@
 import { HOST } from '../../env.js'
+import { argsValidation } from '../../utilites/args-validation.js'
 
 export async function deleteBucketOption() {
     const bucketName = process.argv[4]
-    console.log(bucketName)
+    if (!argsValidation([bucketName])) { return }
+
     const res = await fetch(`${HOST}/bucket/delete?bucketName=${bucketName}`, {
         method: 'DELETE'
     })
